Refactor the function handleChangeOnDates
